+++ package.json
@@ -21,7 +21,6 @@
     "is-url": "1.2.4",
     "jwt-decode": "4.0.0",
     "lodash.get": "4.4.2",
-    "moment": "^2.30.1",
     "pizzip": "3.1.6",
     "react": "18.2.0",
     "react-custom-scrollbars": "4.2.1",
