+++ src/app/ContestSite/PlaybackPage.tsx
@@ -54,7 +54,6 @@ const PlaybackPage: React.FC<ContestProps> = ({ mode, user }) => {
         contest_id: Contest_id,
       },
     });
-  console.log(contestSwitchData);
 
   useEffect(() => {
     if (contestSwitchError) {
