+++ src/graphql/contest/round.graphql
@@ -1,9 +1,9 @@
 # 主要操作对象：contest_round表
 
 # 增
-mutation AddContestRound($contest_id: uuid!, $name: String!, $maps: String) {
+mutation AddContestRound($contest_id: uuid!, $name: String!, $map_id: uuid) {
   insert_contest_round_one(
-    object: { contest_id: $contest_id, name: $name, maps: $maps }
+    object: { contest_id: $contest_id, name: $name, map_id: $map_id }
   ) {
     round_id
   }
@@ -31,6 +31,6 @@ query GetContestRounds($contest_id: uuid!) {
   contest_round(where: { contest_id: { _eq: $contest_id } }) {
     round_id
     name
-    maps
+    map_id
   }
 }
